Restructured convert package for easier testability and readability.
Using tinygo regular go tests won't work because the machine package isn't in GOROOT. The work around is to place all functions you wish to test in seperate packages that don't require the machine import.

+++ types/types.go
@@ -10,7 +10,7 @@ type SensorReadings struct {
 }
 
 type TippingBucket struct {
-	Inches float32 `json:"inches"`
+	Inches float64 `json:"inches"`
 	Milimeteres float32 `json:"milimeters"`
 }
